removed old deprecated 'extended_packet_size'
